Refactor data inspection feature
